Extract multi_buffer module out of editor
