+++ crates/multi_buffer/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "editor"
+name = "multi_buffer"
 version = "0.1.0"
 edition = "2021"
 publish = false
 
 [lib]
-path = "src/editor.rs"
+path = "src/multi_buffer.rs"
 doctest = false
 
 [features]
@@ -14,10 +14,7 @@ test-support = [
     "text/test-support",
     "language/test-support",
     "gpui/test-support",
-    "multi_buffer/test-support",
-    "project/test-support",
     "util/test-support",
-    "workspace/test-support",
     "tree-sitter-rust",
     "tree-sitter-typescript"
 ]
@@ -25,19 +22,12 @@ test-support = [
 [dependencies]
 client = { path = "../client" }
 clock = { path = "../clock" }
-copilot = { path = "../copilot" }
-db = { path = "../db" }
-drag_and_drop = { path = "../drag_and_drop" }
 collections = { path = "../collections" }
 context_menu = { path = "../context_menu" }
-fuzzy = { path = "../fuzzy" }
 git = { path = "../git" }
 gpui = { path = "../gpui" }
 language = { path = "../language" }
 lsp = { path = "../lsp" }
-multi_buffer = { path = "../multi_buffer" }
-project = { path = "../project" }
-rpc = { path = "../rpc" }
 rich_text = { path = "../rich_text" }
 settings = { path = "../settings" }
 snippet = { path = "../snippet" }
@@ -45,8 +35,6 @@ sum_tree = { path = "../sum_tree" }
 text = { path = "../text" }
 theme = { path = "../theme" }
 util = { path = "../util" }
-sqlez = { path = "../sqlez" }
-workspace = { path = "../workspace" }
 
 aho-corasick = "1.1"
 anyhow.workspace = true
@@ -59,6 +47,7 @@ log.workspace = true
 ordered-float.workspace = true
 parking_lot.workspace = true
 postage.workspace = true
+pulldown-cmark = { version = "0.9.2", default-features = false }
 rand.workspace = true
 schemars.workspace = true
 serde.workspace = true
